Move other contents from execution to batch-spec/execute, reduce mock data sources
