+++ client/web/src/enterprise/batches/batch-spec/execute/ExecuteBatchSpecPage.tsx
@@ -26,12 +26,12 @@ import {
 // TODO: Make sure fields on GraphQL queries are all still needed
 import { GET_BATCH_CHANGE_TO_EDIT } from '../../create/backend'
 import { ConfigurationForm } from '../../create/ConfigurationForm'
-import { FETCH_BATCH_SPEC_EXECUTION } from '../../execution/backend'
 import { NewBatchChangePreviewPage } from '../../preview/BatchChangePreviewPage'
 import { BatchSpecContext, BatchSpecContextProvider } from '../BatchSpecContext'
 import { BatchChangeHeader } from '../header/BatchChangeHeader'
 import { TabBar, TabsConfig } from '../TabBar'
 
+import { FETCH_BATCH_SPEC_EXECUTION } from './backend'
 import { ExecutionWorkspaces } from './ExecutionWorkspaces'
 import { ReadOnlyBatchSpecForm } from './ReadOnlyBatchSpecForm'
 
@@ -117,6 +117,18 @@ const ExecuteBatchSpecPageContent: React.FunctionComponent<
         [batchSpec.applyURL]
     )
 
+    const description =
+        batchSpec.contextType === 'full' ? (
+            <>
+                Created <Timestamp date={batchSpec.createdAt} /> by{' '}
+                <LinkOrSpan to={batchSpec.creator?.url}>
+                    {batchSpec.creator?.displayName || batchSpec.creator?.username || 'a deleted user'}
+                </LinkOrSpan>
+            </>
+        ) : (
+            batchChange.description
+        )
+
     return (
         <div className={layoutStyles.pageContainer}>
             <div className={layoutStyles.headerContainer}>
@@ -126,14 +138,7 @@ const ExecuteBatchSpecPageContent: React.FunctionComponent<
                         text: batchChange.namespace.namespaceName,
                     }}
                     title={{ to: batchChange.url, text: batchChange.name }}
-                    description={
-                        <>
-                            Created <Timestamp date={batchSpec.createdAt} /> by{' '}
-                            <LinkOrSpan to={batchSpec.creator?.url}>
-                                {batchSpec.creator?.displayName || batchSpec.creator?.username || 'a deleted user'}
-                            </LinkOrSpan>
-                        </>
-                    }
+                    description={description}
                 />
             </div>
