+++ client/web/src/enterprise/batches/batch-spec/execute/util.ts
@@ -1,4 +1,4 @@
-import { BatchSpecWorkspaceState } from '../../../graphql-operations'
+import { BatchSpecWorkspaceState } from '../../../../graphql-operations'
 
 export function isValidBatchSpecWorkspaceState(state: string): state is BatchSpecWorkspaceState {
     return Object.values(BatchSpecWorkspaceState).some(value => value === state)
