+++ client/web/src/enterprise/batches/batch-spec/header/ExecutionStats.tsx
@@ -1,0 +1,1 @@
+export const ExecutionStats: React.FunctionComponent<{}> = () => <div>I'm execution stats!</div>
